# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export type OpenaiMessageMetadata = { [key: string]: unknown } | null;

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  metadata?: OpenaiMessageMetadata;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type DraftOrderItemType = typeof DraftOrderItemType[keyof typeof DraftOrderItemType];


export const DraftOrderItemType = {
  shopify: 'shopify',
  external: 'external',
} as const;

export interface DraftOrderItem {
  type: DraftOrderItemType;
  /** @nullable */
  variantId?: string | null;
  title: string;
  price: number;
  quantity: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface DraftOrderInput {
  items: DraftOrderItem[];
  note?: string;
}

export interface DraftOrderResult {
  checkoutUrl: string;
  draftOrderId: string;
}