# lib/api-zod/src/generated/types/draftOrderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DraftOrderItem } from './draftOrderItem';

export interface DraftOrderInput {
  items: DraftOrderItem[];
  note?: string;
}