# lib/api-zod/src/generated/types/draftOrderItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DraftOrderItemType } from './draftOrderItemType';

export interface DraftOrderItem {
  type: DraftOrderItemType;
  /** @nullable */
  variantId?: string | null;
  title: string;
  price: number;
  quantity: number;
  /** @nullable */
  imageUrl?: string | null;
}