# lib/api-zod/src/generated/types/openaiMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OpenaiMessageMetadata } from './openaiMessageMetadata';

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  metadata?: OpenaiMessageMetadata;
  createdAt: Date;
}